kafka version bumped to 0.9.0.1; createTopic() API had changed, TopicCommand.main() exits now, fixed by calling TopicCommand.createTopic() instead; documenting kafka versions; test timeout extended to prevent timing issues

+++ src/test/java/info/batey/kafka/unit/KafkaIntegrationTest.java
@@ -46,7 +46,7 @@ public class KafkaIntegrationTest {
         Field f = kafkaUnitServer.getClass().getDeclaredField("broker");
         f.setAccessible(true);
         KafkaServerStartable broker = (KafkaServerStartable) f.get(kafkaUnitServer);
-        assertEquals(1024, broker.serverConfig().logSegmentBytes());
+        assertEquals(1024, (int)broker.serverConfig().logSegmentBytes());
 
         kafkaUnitServer.shutdown();
     }
